requirements.txt
structlog==23.1.0
blake3==0.3.3
httpx==0.25.0
pytest>=8.2,<9
pypdf==4.3.1
textract==1.6.3
pytesseract==0.3.10
Pillow==10.0.1
mammoth==1.6.0
html2text==2020.1.16
python-docx==1.1.0
easyocr==1.7.1
